Created a hook for fetching directories. moved server hooks into their own folder
